requirements.txt
ml-collections==1.1.0
torch==2.10.0
torchvision==0.25.0
pillow>=10.0.0
scipy>=1.17.0
numpy>=1.26.0
huggingface-hub>=0.23.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0